% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_distributions.R
\name{clear.independence}
\alias{clear.independence}
\title{Clear Independence Assumptions}
\usage{
clear.independence()
}
\description{
Removes all independence assumptions.
}
\examples{
assume.independent(X, Y)
clear.independence()
E(X * Y)  # Returns: E(X * Y) (no longer factors)
}
