% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{clear_template_cache}
\alias{clear_template_cache}
\title{Clear the template cache}
\usage{
clear_template_cache()
}
\description{
Internal helper. Useful in tests that modify templates between calls.
}
\keyword{internal}
