% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{combine.powers}
\alias{combine.powers}
\title{Combine Powers}
\usage{
combine.powers(left, right)
}
\arguments{
\item{left}{Left expression}

\item{right}{Right expression}
}
\value{
Combined power expression, or NULL if not combinable
}
\description{
Combines power expressions: x\emph{x^n = x^(n+1), x^m}x^n = x^(m+n)
}
\keyword{internal}
