% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_user_wrappers.R
\name{Cov}
\alias{Cov}
\title{Symbolic Covariance}
\usage{
Cov(x, y)
}
\arguments{
\item{x}{First random variable expression}

\item{y}{Second random variable expression}
}
\value{
An unevaluated expression representing Cov\link{x, y}
}
\description{
Computes the symbolic covariance using the identity
Cov(X, Y) = E(XY) - E(X)E(Y)
}
\examples{
Cov(X, Y)      # Returns: E(X * Y) - E(X) * E(Y)
}
