% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_latex.R
\name{expr.to.latex}
\alias{expr.to.latex}
\title{Expression to LaTeX (Recursive)}
\usage{
expr.to.latex(expr)
}
\arguments{
\item{expr}{An R expression}
}
\value{
LaTeX string
}
\description{
Recursively converts an R expression tree to LaTeX.
}
\keyword{internal}
