% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{formula_lhs}
\alias{formula_lhs}
\title{Get the left-hand side of a formula}
\usage{
formula_lhs(f)
}
\arguments{
\item{f}{A formula.}
}
\value{
A language object or \code{NULL} for one-sided formulas.
}
\description{
Get the left-hand side of a formula
}
\keyword{internal}
