% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{formula_rhs}
\alias{formula_rhs}
\title{Get the right-hand side of a formula}
\usage{
formula_rhs(f)
}
\arguments{
\item{f}{A formula.}
}
\value{
A language object.
}
\description{
Get the right-hand side of a formula
}
\keyword{internal}
