% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_internal_simplify.R
\name{normalize.base}
\alias{normalize.base}
\title{Normalize Base Expression}
\usage{
normalize.base(base)
}
\arguments{
\item{base}{A base expression}
}
\value{
A character string key
}
\description{
Creates a canonical string representation of a base expression
for grouping like terms.
}
\keyword{internal}
