% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_distributions.R
\name{print.conditional_moment}
\alias{print.conditional_moment}
\title{Print Conditional Moment}
\usage{
\method{print}{conditional_moment}(x, ...)
}
\arguments{
\item{x}{A conditional moment object}

\item{...}{Additional arguments}
}
\description{
Custom print method for conditional expressions.
Formats 'when(expr, cond)' as '\link{expr} when \link{cond}'.
}
