% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_calculus.R
\name{restore.E}
\alias{restore.E}
\title{Restore E() Calls After Differentiation}
\usage{
restore.E(expr, map)
}
\arguments{
\item{expr}{An expression with \emph{\emph{E}...}_ symbols}

\item{map}{The substitution map from protect.E}
}
\value{
The expression with E() calls restored
}
\description{
Replaces temporary symbols back to their original E() calls.
}
\keyword{internal}
