% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_calculus.R
\name{substitute.var}
\alias{substitute.var}
\title{Substitute Variable with Value}
\usage{
substitute.var(expr, var, val)
}
\arguments{
\item{expr}{The expression}

\item{var}{The variable to replace (symbol)}

\item{val}{The value to substitute}
}
\value{
Modified expression
}
\description{
Replaces all occurrences of a variable with a value.
}
\keyword{internal}
