% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_latex.R
\name{symbol.to.latex}
\alias{symbol.to.latex}
\title{Convert Symbol to LaTeX}
\usage{
symbol.to.latex(name)
}
\arguments{
\item{name}{Character name of the symbol}
}
\value{
LaTeX representation
}
\description{
Handles Greek letter conversion and subscript notation.
}
\keyword{internal}
