% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolizer-package.R
\docType{package}
\name{symbolizer-package}
\alias{symbolizer}
\alias{symbolizer-package}
\title{symbolizer: Structured Symbolic Specifications, Interpretations, and Teachable Stories for Modern Statistical Models}
\description{
Turns fitted distributional, multivariate, latent-variable, and spatial regression models into structured symbolic specifications that render as publication-ready LaTeX equations, assumption tables, formula bridges, parameter interpretations, and (later) structural model comparisons, methods text, simulation recipes, and model graphs. Companion to 'drmTMB' and 'gllvmTMB'; complementary to 'equatiomatic' for one-formula models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://itchyshin.github.io/symbolizer/}
  \item \url{https://github.com/itchyshin/symbolizer}
  \item Report bugs at \url{https://github.com/itchyshin/symbolizer/issues}
}

}
\author{
\strong{Maintainer}: Shinichi Nakagawa \email{itchyshin@gmail.com} (\href{https://orcid.org/0000-0002-7765-5182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Shinichi Nakagawa \email{itchyshin@gmail.com} (\href{https://orcid.org/0000-0002-7765-5182}{ORCID}) [copyright holder]
}

}
\keyword{internal}
