% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{tag.conditional}
\alias{tag.conditional}
\title{Tag Conditional Expression}
\usage{
tag.conditional(expr)
}
\arguments{
\item{expr}{An expression}
}
\value{
The expression, possibly with an added class
}
\description{
Checks if an expression is a 'when' call and tags it with 'conditional_moment' class.
}
\keyword{internal}
