% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_latex.R
\name{to.latex.default}
\alias{to.latex.default}
\title{Default LaTeX Conversion}
\usage{
\method{to.latex}{default}(x, delimiters = "none", ...)
}
\arguments{
\item{x}{An R expression or call object}

\item{delimiters}{Character; "none", "inline" ($), or "display" (\[\])}

\item{...}{Additional arguments (ignored)}
}
\value{
A latex_output object containing LaTeX code
}
\description{
Converts R expressions to LaTeX using expression tree traversal.
}
