% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_internal_simplify.R
\name{try.combine.like.terms}
\alias{try.combine.like.terms}
\title{Try to Combine Like Terms}
\usage{
try.combine.like.terms(left, right, op)
}
\arguments{
\item{left}{Left simplified expression}

\item{right}{Right simplified expression}

\item{op}{The operator ("+" or "-")}
}
\value{
Combined expression, or NULL if terms are not alike
}
\description{
Attempts to combine like terms in addition/subtraction.
a\emph{X + b}X -> (a+b)\emph{X, a}X - b*X -> (a-b)*X
}
\keyword{internal}
