% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolizer-package.R
\docType{package}
\name{symbolizer-package}
\alias{symbolizer}
\alias{symbolizer-package}
\title{symbolizer: Symbolic Specifications, Interpretations, and Teachable Stories for Statistical Models}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Turns fitted distributional, multivariate, latent-variable, spatial, mixed-effect, meta-analytic, and additive models into structured symbolic specifications that render as publication-ready LaTeX equations, assumption tables, formula bridges, parameter interpretations, structural model comparisons, methods-section paragraphs, and structural DAGs. First-class methods for 'drmTMB', 'gllvmTMB', 'glmmTMB', 'brms', 'lme4', 'MCMCglmm', 'sdmTMB', 'stats::lm' / 'stats::glm', 'metafor', and 'mgcv'. Complementary to 'equatiomatic' for one-formula models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://itchyshin.github.io/symbolizer/}
  \item \url{https://github.com/itchyshin/symbolizer}
  \item Report bugs at \url{https://github.com/itchyshin/symbolizer/issues}
}

}
\author{
\strong{Maintainer}: Shinichi Nakagawa \email{itchyshin@gmail.com} (\href{https://orcid.org/0000-0002-7765-5182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Shinichi Nakagawa \email{itchyshin@gmail.com} (\href{https://orcid.org/0000-0002-7765-5182}{ORCID}) [copyright holder]
}

}
\keyword{internal}
