% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{wrap_aligned}
\alias{wrap_aligned}
\title{Wrap a vector of equation lines in \code{aligned} environment}
\usage{
wrap_aligned(lines)
}
\arguments{
\item{lines}{Character vector of LaTeX equation lines.}
}
\value{
A single character string: \code{lines} joined with \verb{\\} line breaks and
wrapped in a LaTeX \code{aligned} environment.
}
\description{
Wrap a vector of equation lines in \code{aligned} environment
}
\keyword{internal}
